Count homepage items
