+++ src/index.html
@@ -18,8 +18,8 @@
   <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <div class="container">
       <a class="navbar-brand" href="index.html">FoodBox</a>
-      <p class="mt-3">Item Count : <span id="itemCount"></span></p>
-      <div id="meal-dish-counter"></div>
+      <h5 class="mt-2">Dishes (<span id="itemCount"></span>)</h5>
+      <h2 id="meal-dish-counter"></h2>
       <div id="comments-counter"></div>
       <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
         aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
