+++ src/index.js
@@ -2,13 +2,9 @@ import './style.css';
 
 import { fetchData, getAppId } from './modules/app.js';
 
-import mealDishesCounter from './modules/itemCtr.js';
+import mealDishesCounter from './modules/counter/itemCtr.js';
 
-import commentsCounter from './modules/commentsCtr.js';
-
-// Call the countMealDishes function to display the total number of meal dishes on the homepage
-// mealDishesCounter.countMealDishes();
-// commentsCounter.countComments();
+import commentsCounter from './modules/counter/commentsCtr.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   fetchData();
