+++ src/modules/app.js
@@ -117,7 +117,7 @@ export const fetchData = async () => {
       textarea.style = 'resize:none';
 
       const button2 = document.createElement('button');
-      button2.classList = 'btn btn-primary mb-3 float-end';
+      button2.classList = 'btn btn-outline-info mb-3 float-end';
       button2.innerHTML = 'Comment';
 
       div1.appendChild(div2);
