+++ src/modules/category.js
@@ -26,7 +26,7 @@ const renderCategory = (category) => {
 
   // Create a like button for this category
   const likeButton = document.createElement('button');
-  likeButton.className = 'btn btn-outline-danger like-button my-like-button';
+  likeButton.className = 'btn btn-outline-danger like-button my-like-button my-2';
   likeButton.dataset.categoryId = category.idCategory;
   likeButton.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M12 21.35l-1.45-1.32C5.4 14.35 2 11.12 2 7.5 2 4.42 4.42 2 7.5 2c1.93 0 3.75.95 4.87 2.53C13.75 2.95 15.57 2 17.5 2 20.58 2 23 4.42 23 7.5c0 3.62-3.4 6.85-8.55 12.53L12 21.35z"/></svg>';
